# Faster BooleanBufferBuilder::append_n for true values