# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

Inspired by https://github.com/apache/arrow-rs/pull/1603. The performance benefits are clearly visible when decoding nullable primitive values from parquet, where it yields a non-trivial speedup

```
arrow_array_reader/Int64Array/binary packed, mandatory, no NULLs
                        time:   [17.706 µs 17.711 µs 17.718 µs]
                        change: [+6.7854% +6.8707% +6.9525%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
arrow_array_reader/Int64Array/binary packed, optional, no NULLs
                        time:   [18.660 µs 18.678 µs 18.710 µs]
                        change: [-37.996% -37.947% -37.880%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
arrow_array_reader/Int64Array/binary packed skip, mandatory, no NULLs
                        time:   [16.585 µs 16.595 µs 16.608 µs]
                        change: [+5.0752% +5.1957% +5.3004%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
arrow_array_reader/Int64Array/binary packed skip, optional, no NULLs
                        time:   [17.200 µs 17.207 µs 17.214 µs]
                        change: [-27.663% -27.598% -27.528%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
arrow_array_reader/Int64Array/binary packed, optional, half NULLs
                        time:   [29.343 µs 29.349 µs 29.356 µs]
                        change: [+4.4487% +4.5826% +4.6903%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe
```

Note the regressions are just noise, they fluctuate by ~8%.

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?

No

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->